+++ example/lib/routes/components/slider.dart
@@ -205,8 +205,8 @@ class _TSliderCustomBounds extends StatelessWidget {
       max: 10000,
       initialValue: 5500,
       showDefaultMarks: true,
-      formatter: (value) => value.toInt().autoFormat(),
-      tooltipFormatter: (value) => value.toInt().autoFormat(),
+      formatter: (value) => value.toInt().toDisplayString(),
+      tooltipFormatter: (value) => value.toInt().toDisplayString(),
       onChanged: (value) {},
     );
   }